+++ src/geopub/db.cljs
@@ -2,8 +2,9 @@
   (:require [re-frame.core :as re-frame]
             [rdf.core :as rdf]
             [rdf.graph.map]
+            [rdf.fragment-graph]
             [cljs.core.async :as async :refer [<!]])
-  (:require-macros [cljs.core.async :refer [go]]))
+  (:require-macros [cljs.core.async :refer [go-loop]]))
 
 (re-frame/reg-event-db
  ::initialize
@@ -21,10 +22,14 @@
   (satisfies? cljs.core.async.impl.protocols/Channel x))
 
 
-;; a graph in an async channel needs to be handled as a side effect
+;; take graphs from channel and individually add to db
 (re-frame/reg-fx
  ::add-rdf-graph-async
- (fn [chan] (go (re-frame/dispatch [::add-rdf-graph (<! chan)]))))
+ (fn [chan]
+   (go-loop []
+     (when-let [graph (<! chan)]
+       (re-frame/dispatch [::add-rdf-graph graph])
+       (recur)))))
 
 (re-frame/reg-event-fx
  ::add-rdf-graph
@@ -34,7 +39,7 @@
      ;; if an RDF graph add to database
      (rdf/graph? graph) {:db (rdf/graph-merge (:db coeffects) graph)}
 
-     ;; if a channel take first element from channel and retry
+     ;; handle async/channel of graphs
      (channel? graph) {::add-rdf-graph-async graph}
 
      :else {})))