+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [UNRELEASED]
 
+- Authentication with OAuth 2.0
+- Refactor to [re-frame framework](https://day8.github.io/re-frame/re-frame/)
+- Support [content-addressable RDF](https://openengiadina.net/papers/content-addressable-rdf.html)
+- [ERIS encoding](https://openengiadina.net/papers/eris.html) for content-addressing
 
 ## [0.1.0] - 2020-04-03